requirements.txt
Django==2.2.13
django-widget-tweaks==1.4.3

django-bootstrap-modal-forms==2.2.0

# For image fields and form generator
Pillow==8.2.0

langdetect==1.0.9
openpyxl==3.0.7

# Django global login required middleware
django-glrm==1.1.3

jdatetime==3.6.2

# date time localization
babel==2.8.0

django-polymorphic